bash/aliases: systemctl: check whether systemd is running too
For Debian where systemd is installed but not yet in use.

+++ bash/aliases.sh
@@ -347,7 +347,7 @@ owns() {
 
 ## service management
 
-if have systemctl; then
+if have systemctl && [ -d /run/systemd/system ]; then
 	start()   { sudo systemctl start "$@"; systemctl status -a "$@"; }
 	stop()    { sudo systemctl stop "$@"; systemctl status -a "$@"; }
 	restart() { sudo systemctl restart "$@"; systemctl status -a "$@"; }
